修改部分细节问题

+++ src/main/java/cn/wzy/Listener.java
@@ -5,16 +5,16 @@ import org.apache.kafka.clients.producer.RecordMetadata;
 import org.springframework.kafka.support.ProducerListener;
 
 
-public class Listener implements ProducerListener<Integer, String> {
+public class Listener implements ProducerListener<Long, String> {
 
-  public void onSuccess(String topic, Integer partition, Integer key, String value, RecordMetadata recordMetadata) {
+  public void onSuccess(String topic, Integer partition, Long key, String value, RecordMetadata recordMetadata) {
     System.out.println("====success====");
     System.out.println("topic " + topic);
     System.out.println("====success====");
   }
 
   @Override
-  public void onError(ProducerRecord<Integer, String> producerRecord, Exception exception) {
+  public void onError(ProducerRecord<Long, String> producerRecord, Exception exception) {
     exception.printStackTrace();
   }
 }
